+++ fabric/src/main/java/ellemes/expandedstorage/fabric/mixin/HTMChestCompat.java
@@ -1,4 +1,4 @@
-package ellemes.expandedstorage.mixin;
+package ellemes.expandedstorage.fabric.mixin;
 
 import com.github.fabricservertools.htm.HTMContainerLock;
 import com.github.fabricservertools.htm.api.LockableChestBlock;