+++ quilt/src/main/java/ellemes/expandedstorage/quilt/mixin/HTMChestCompat.java
@@ -1,0 +1,29 @@
+package ellemes.expandedstorage.quilt.mixin;
+
+import com.github.fabricservertools.htm.HTMContainerLock;
+import com.github.fabricservertools.htm.api.LockableChestBlock;
+import ellemes.expandedstorage.block.AbstractChestBlock;
+import ellemes.expandedstorage.thread.compat.htm.HTMChestProperties;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
+import org.spongepowered.asm.mixin.Mixin;
+
+import java.util.Optional;
+
+@Mixin(AbstractChestBlock.class)
+public abstract class HTMChestCompat implements LockableChestBlock {
+    @Override
+    public HTMContainerLock getLockAt(BlockState state, Level world, BlockPos pos) {
+        //noinspection ConstantConditions
+        return AbstractChestBlock.createPropertyRetriever((AbstractChestBlock) (Object) this, state, world, pos, true).get(HTMChestProperties.LOCK_PROPERTY).orElse(null);
+    }
+
+    @Override
+    public Optional<BlockEntity> getUnlockedPart(BlockState state, Level world, BlockPos pos) {
+        //noinspection ConstantConditions
+        return AbstractChestBlock.createPropertyRetriever((AbstractChestBlock) (Object) this, state, world, pos, true).get(HTMChestProperties.UNLOCKED_BE_PROPERTY);
+    }
+
+}