+++ quilt/src/main/java/ellemes/expandedstorage/quilt/mixin/HTMLockableBlockEntityCompat.java
@@ -1,4 +1,4 @@
-package ellemes.expandedstorage.mixin;
+package ellemes.expandedstorage.quilt.mixin;
 
 import com.github.fabricservertools.htm.HTMContainerLock;
 import com.github.fabricservertools.htm.api.LockableObject;