+++ changelog.md
@@ -1,3 +1,1 @@
-- Fix up some metadata in files.
-- Update forge to 41.0.64 ( breaking changes ).
-  - Thanks to XFactHD for helping with rendering changes.
+- Fix mixin issues in fabric and quilt projects.